+++ sphinx_needs/api/need.py
@@ -196,7 +196,11 @@ def add_need(
     else:
         need_id = id
 
-    if needs_config.id_regex and not re.match(needs_config.id_regex, need_id):
+    if (
+        needs_config.id_regex
+        and not is_external
+        and not re.match(needs_config.id_regex, need_id)
+    ):
         raise NeedsInvalidException(
             f"Given ID '{need_id}' does not match configured regex '{needs_config.id_regex}'"
         )